minor logic change + refactoring
